requirements.txt
Flask==2.0.3
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.0.1
tensorflow==2.8.0
tf_nightly==2.9.0.dev20220222
